requirements.txt
quart
disnake
discord.py
PyDrive